Add files via upload
This program starts by asking username and password of manager and waiter, the user can change from one waiter to another waiter and the orders will not be deleted. When user exchange account from waiter1 to waiter2 the program doesnt close their activity they are still active. There is a time that calculates the wage by the hour so when a waiter logs out (requiring manager password) it will show tables served, day profit, working hours and wage earned.